add safety net for bad artifact id per #14
There have been a couple of instances where the Rally service
has returned a `null` instead of a valid work item reference,
which results in a thrown `RallyAssetNotFoundException` and a
failing build. This implementation short-circuits the commit
creation logic in that case, essentially swallowing the
exception and allowing execution to continue unhindered.